Add Service Layer and adjusted class attr
The StockService will rely on the new StockRepository and OptionRepository to save class instances. A stones throw from being able to save a single ticker with current options data into our DB!! Then we can run some calculations and package it up for js. Subject to change again, but I edited a couple of attributes since the first commit, they should be non-issue, but if you disagree, please say so.