Skeleton Test API call
New API has 800/calls a month and we use it to call the ticker to retrieve a list of dates of option expiries. Next we'll use that array to make calls to this new api endpoint using the ticker symbol & the expiry to get sets of options data. Almost there!