Demo, graph with outside data rendered on next page
Includes a couple of new paths that are for testing purposes. The buttons are linked in without new make api calls. will render your last AAPL update in the DB for displaying purposes. The search from the home page still operates normally, as does the response and view on options.jsp.