Adapting fetchStockData for error handling
The Mboum API doesn't return an error response when searching four letter combinations that don't correspond to active tickers. The response for a bad request is an unfinished shell of an object, with an empty array where results would be. Because this still passes http status 200, return null to redirect in controller validation.